+++ classes/suggested-tasks/providers/integrations/yoast/class-media-pages.php
@@ -20,6 +20,13 @@ class Media_Pages extends Yoast_Provider {
 	protected const PROVIDER_ID = 'yoast-media-pages';
 
 	/**
+	 * The external link URL.
+	 *
+	 * @var string
+	 */
+	protected const EXTERNAL_LINK_URL = 'https://prpl.fyi/yoast-media-pages';
+
+	/**
 	 * Get the task URL.
 	 *
 	 * @return string
@@ -43,11 +50,7 @@ class Media_Pages extends Yoast_Provider {
 	 * @return string
 	 */
 	protected function get_description() {
-		return \sprintf(
-			/* translators: %s: "Read more" link. */
-			\esc_html__( 'Yoast SEO can disable the media / attachment pages, which are the pages that show the media files. You really don\'t need them, except when you are displaying photos or art on your site through them. %s.', 'progress-planner' ),
-			'<a href="https://prpl.fyi/yoast-media-pages" target="_blank" data-prpl_accessibility_text="' . \esc_attr__( 'Read more about the Yoast SEO Media Pages', 'progress-planner' ) . '">' . \esc_html__( 'Read more', 'progress-planner' ) . '</a>'
-		);
+		return \esc_html__( 'Yoast SEO can disable the media / attachment pages, which are the pages that show the media files. You really don\'t need them, except when you are displaying photos or art on your site through them.', 'progress-planner' );
 	}
 
 	/**
@@ -77,5 +80,22 @@ class Media_Pages extends Yoast_Provider {
 	public function should_add_task() {
 		// If the media pages are already disabled, we don't need to add the task.
 		return \YoastSEO()->helpers->options->get( 'disable-attachment' ) !== true;
+	}
+
+	/**
+	 * Add task actions specific to this task.
+	 *
+	 * @param array $data    The task data.
+	 * @param array $actions The existing actions.
+	 *
+	 * @return array
+	 */
+	public function add_task_actions( $data = [], $actions = [] ) {
+		$actions[] = [
+			'priority' => 10,
+			'html'     => '<a class="prpl-tooltip-action-text" href="' . \admin_url( 'admin.php?page=wpseo_page_settings#/media-pages' ) . '" target="_blank">' . \esc_html__( 'Disable', 'progress-planner' ) . '</a>',
+		];
+
+		return $actions;
 	}
 }