+++ CHANGELOG.md
@@ -2,8 +2,12 @@
 = 1.8.0 =
 
 Added these recommendations from Ravi:
+
 * Set date format.
 
+Enhancements:
+
+* Redesign of the admin pages.
 
 = 1.7.2 =